Fix MSVC warnings and errors: min/max macros, localtime, size_t conversions

+++ tui.cpp
@@ -600,9 +600,9 @@ void TUI::draw_status_bar(std::stringstream& output, int cols) {
                          " | Total: " + std::to_string(proxy_server_->get_total_connections());
     
     // Calculate actual visible width (ANSI codes don't count)
-    int title_len = title.length();
-    int status_len = status_text.length();
-    int metrics_len = metrics.length();
+    int title_len = static_cast<int>(title.length());
+    int status_len = static_cast<int>(status_text.length());
+    int metrics_len = static_cast<int>(metrics.length());
     int space_len = 1; // Space between status and metrics
     
     int total_content_len = title_len + status_len + metrics_len + space_len;
@@ -696,7 +696,7 @@ void TUI::draw_content_area(std::stringstream& output, int cols, int max_rows) {
 // Table rendering helpers
 void TUI::draw_table_border(std::stringstream& output, const std::string& title, int cols) {
     output << "┌─ " << title;
-    int used = 3 + title.length();
+    int used = 3 + static_cast<int>(title.length());
     for (int i = used; i < cols - 1; ++i) {
         output << "─";
     }
@@ -709,7 +709,7 @@ void TUI::draw_table_header(std::stringstream& output, const std::vector<std::pa
         std::string header = col.first;
         int width = col.second;
         output << " " << header;
-        int padding = width - header.length() - 1;
+        int padding = width - static_cast<int>(header.length()) - 1;
         for (int i = 0; i < padding; ++i) {
             output << " ";
         }
@@ -767,7 +767,7 @@ void TUI::draw_table_row(std::stringstream& output, const std::vector<std::strin
         }
         
         output << " " << cell;
-        int padding = width - cell.length() - 1;
+        int padding = width - static_cast<int>(cell.length()) - 1;
         for (int j = 0; j < padding; ++j) {
             output << " ";
         }
@@ -1187,7 +1187,7 @@ void TUI::draw_help_tab(std::stringstream& output, int cols, int max_rows UNUSED
     for (const auto& shortcut : shortcuts) {
         output << "│ " << std::setw(12) << std::left << shortcut.first;
         output << " " << shortcut.second;
-        int used = 15 + shortcut.first.length() + shortcut.second.length();
+        int used = 15 + static_cast<int>(shortcut.first.length() + shortcut.second.length());
         for (int i = used; i < cols - 1; ++i) output << " ";
         output << "│\n";
     }
@@ -1213,7 +1213,7 @@ void TUI::draw_help_tab(std::stringstream& output, int cols, int max_rows UNUSED
     for (const auto& op : mouse_ops) {
         output << "│ " << std::setw(15) << std::left << op.first;
         output << " " << op.second;
-        int used = 18 + op.first.length() + op.second.length();
+        int used = 18 + static_cast<int>(op.first.length() + op.second.length());
         for (int i = used; i < cols - 1; ++i) output << " ";
         output << "│\n";
     }
@@ -1247,8 +1247,8 @@ void TUI::draw_summary_bar(std::stringstream& output, int cols) {
     std::string separator = " | ";
     
     // Calculate visible length (3 separators between 4 items)
-    int visible_len = stats_label.length() + runways_text.length() + targets_text.length() + 
-                      conns_text.length() + throughput_text.length() + (3 * separator.length());
+    int visible_len = static_cast<int>(stats_label.length() + runways_text.length() + targets_text.length() + 
+                      conns_text.length() + throughput_text.length() + (3 * separator.length()));
     
     // Output with ANSI codes for bold "Stats:"
     output << "\033[1m" << stats_label << "\033[0m";
@@ -1276,7 +1276,7 @@ void TUI::draw_command_bar(std::stringstream& output, int cols) {
     }
     
     output << command_text;
-    int used = command_text.length();
+    int used = static_cast<int>(command_text.length());
     
     // Fill remaining space with background color
     fill_line_with_bg(output, used, cols, bg_color);
@@ -1290,7 +1290,7 @@ void TUI::draw_detail_view(std::stringstream& output, int cols, int rows) {
     output << bg_color;
     std::string title = " Detail View ";
     output << title;
-    int used = title.length();
+    int used = static_cast<int>(title.length());
     
     // Fill remaining space with background color
     fill_line_with_bg(output, used, cols, bg_color);
